Cleaned up HostProxy, removing the duplicate data between HostProxy and the outer Func object. Also updated the reflection generator to be able to generate delegate code.
