+++ dergwasm_mod/Benchmarks/Program.cs
@@ -4,6 +4,7 @@ using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Running;
 using Derg;
+using Derg.Modules;
 using Derg.Wasm;
 using DergwasmTests.instructions;
 using DergwasmTests.testing;
@@ -239,9 +240,10 @@ namespace DergwasmTests
             //
             // Func 14 (= idx 4): host func
             machine.SetProgram(0, I32Const(10), I32Const(20), Call(4), Nop());
+            var hostFunc = ModuleReflector.ReflectHostFunc("test", "env", new Func<int, int, int>((a, b) => a - b));
             machine.SetHostFuncAt(
                 14,
-                new ReturningHostProxy<int, int, int>((Frame f, int a, int b) => a - b)
+                hostFunc.FuncFactory().Proxy
             );
         }
