+++ dergwasm_mod/Dergwasm/Machine.cs
@@ -309,239 +309,6 @@ namespace Derg
             hostFuncs.Add($"{hostFunc.ModuleName}.{hostFunc.Name}", hostFunc);
         }
 
-        public void RegisterHostFunc(
-            string moduleName,
-            string name,
-            FuncType signature,
-            HostProxy proxy
-        )
-        {
-            RegisterHostFunc(new HostFunc(moduleName, name, signature, proxy));
-        }
-
-        public void RegisterVoidHostFunc(string moduleName, string name, Action<Frame> func)
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(new ValueType[] { }, new ValueType[] { }),
-                new VoidHostProxy(func)
-            );
-        }
-
-        public void RegisterVoidHostFunc<T1>(string moduleName, string name, Action<Frame, T1> func)
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(new ValueType[] { Value.ValueType<T1>() }, new ValueType[] { }),
-                new HostProxy<T1>(func)
-            );
-        }
-
-        public void RegisterVoidHostFunc<T1, T2>(
-            string moduleName,
-            string name,
-            Action<Frame, T1, T2> func
-        )
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(
-                    new ValueType[] { Value.ValueType<T1>(), Value.ValueType<T2>() },
-                    new ValueType[] { }
-                ),
-                new HostProxy<T1, T2>(func)
-            );
-        }
-
-        public void RegisterVoidHostFunc<T1, T2, T3>(
-            string moduleName,
-            string name,
-            Action<Frame, T1, T2, T3> func
-        )
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(
-                    new ValueType[]
-                    {
-                        Value.ValueType<T1>(),
-                        Value.ValueType<T2>(),
-                        Value.ValueType<T3>()
-                    },
-                    new ValueType[] { }
-                ),
-                new HostProxy<T1, T2, T3>(func)
-            );
-        }
-
-        public void RegisterVoidHostFunc<T1, T2, T3, T4>(
-            string moduleName,
-            string name,
-            Action<Frame, T1, T2, T3, T4> func
-        )
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(
-                    new ValueType[]
-                    {
-                        Value.ValueType<T1>(),
-                        Value.ValueType<T2>(),
-                        Value.ValueType<T3>(),
-                        Value.ValueType<T4>()
-                    },
-                    new ValueType[] { }
-                ),
-                new HostProxy<T1, T2, T3, T4>(func)
-            );
-        }
-
-        public void RegisterVoidHostFunc<T1, T2, T3, T4, T5>(
-            string moduleName,
-            string name,
-            Action<Frame, T1, T2, T3, T4, T5> func
-        )
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(
-                    new ValueType[]
-                    {
-                        Value.ValueType<T1>(),
-                        Value.ValueType<T2>(),
-                        Value.ValueType<T3>(),
-                        Value.ValueType<T4>(),
-                        Value.ValueType<T5>()
-                    },
-                    new ValueType[] { }
-                ),
-                new HostProxy<T1, T2, T3, T4, T5>(func)
-            );
-        }
-
-        public void RegisterReturningHostFunc<R>(
-            string moduleName,
-            string name,
-            Func<Frame, R> func
-        )
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(new ValueType[] { }, new ValueType[] { Value.ValueType<R>() }),
-                new ReturningVoidHostProxy<R>(func)
-            );
-        }
-
-        public void RegisterReturningHostFunc<T1, R>(
-            string moduleName,
-            string name,
-            Func<Frame, T1, R> func
-        )
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(
-                    new ValueType[] { Value.ValueType<T1>() },
-                    new ValueType[] { Value.ValueType<R>() }
-                ),
-                new ReturningHostProxy<T1, R>(func)
-            );
-        }
-
-        public void RegisterReturningHostFunc<T1, T2, R>(
-            string moduleName,
-            string name,
-            Func<Frame, T1, T2, R> func
-        )
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(
-                    new ValueType[] { Value.ValueType<T1>(), Value.ValueType<T2>() },
-                    new ValueType[] { Value.ValueType<R>() }
-                ),
-                new ReturningHostProxy<T1, T2, R>(func)
-            );
-        }
-
-        public void RegisterReturningHostFunc<T1, T2, T3, R>(
-            string moduleName,
-            string name,
-            Func<Frame, T1, T2, T3, R> func
-        )
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(
-                    new ValueType[]
-                    {
-                        Value.ValueType<T1>(),
-                        Value.ValueType<T2>(),
-                        Value.ValueType<T3>()
-                    },
-                    new ValueType[] { Value.ValueType<R>() }
-                ),
-                new ReturningHostProxy<T1, T2, T3, R>(func)
-            );
-        }
-
-        public void RegisterReturningHostFunc<T1, T2, T3, T4, R>(
-            string moduleName,
-            string name,
-            Func<Frame, T1, T2, T3, T4, R> func
-        )
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(
-                    new ValueType[]
-                    {
-                        Value.ValueType<T1>(),
-                        Value.ValueType<T2>(),
-                        Value.ValueType<T3>(),
-                        Value.ValueType<T4>()
-                    },
-                    new ValueType[] { Value.ValueType<R>() }
-                ),
-                new ReturningHostProxy<T1, T2, T3, T4, R>(func)
-            );
-        }
-
-        public void RegisterReturningHostFunc<T1, T2, T3, T4, T5, R>(
-            string moduleName,
-            string name,
-            Func<Frame, T1, T2, T3, T4, T5, R> func
-        )
-        {
-            RegisterHostFunc(
-                moduleName,
-                name,
-                new FuncType(
-                    new ValueType[]
-                    {
-                        Value.ValueType<T1>(),
-                        Value.ValueType<T2>(),
-                        Value.ValueType<T3>(),
-                        Value.ValueType<T4>(),
-                        Value.ValueType<T5>()
-                    },
-                    new ValueType[] { Value.ValueType<R>() }
-                ),
-                new ReturningHostProxy<T1, T2, T3, T4, T5, R>(func)
-            );
-        }
-
         public int ResolveHostFunc(string moduleName, string name, FuncType signature)
         {
             string key = $"{moduleName}.{name}";
