+++ dergwasm_mod/Dergwasm/Modules/IHostModule.cs
@@ -5,6 +5,6 @@ namespace Derg.Modules
     public interface IHostModule
     {
         List<HostFunc> Functions { get; }
-        List<Api> ApiData { get; }
+        List<ApiFunc> ApiData { get; }
     }
 }
