+++ dergwasm_mod/Dergwasm/Modules/ReflectedModule.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Linq.Expressions;
 using System.Reflection;
 using Elements.Core;
 
@@ -37,12 +36,6 @@ namespace Derg.Modules
             Name = name;
             Generics = generics;
         }
-    }
-
-    internal class ApiData<T>
-    {
-        public Api Data { get; set; }
-        public Func<T, HostFunc> Function { get; set; }
     }
 
     public class ReflectedModule<T> : IHostModule
@@ -89,161 +82,27 @@ namespace Derg.Modules
                         }
                         boundMethod = boundMethod.MakeGenericMethod(attr.Generics);
                     }
-                    ReflectedFuncs.Add(ReflectCallSite(attr, boundMethod, defaultModule));
+                    ReflectedFuncs.Add(ModuleReflector.ReflectHostFunc<T>(attr.Name ?? boundMethod.Name, attr.Module ?? defaultModule, boundMethod));
                 }
             }
-        }
-
-        private static ValueType ValueTypeFor(Type type)
-        {
-            var valueType = (ValueType)
-                typeof(Value)
-                    .GetMethod(nameof(Value.ValueType))
-                    .MakeGenericMethod(type)
-                    .Invoke(null, null);
-            return valueType;
-        }
-
-        private static ApiData<T> ReflectCallSite(
-            ModFnAttribute attr,
-            MethodInfo method,
-            string defaultModule
-        )
-        {
-            ApiData<T> apiData = new ApiData<T>
-            {
-                Data = new Api
-                {
-                    Module = attr.Module ?? defaultModule,
-                    Name = attr.Name ?? method.Name,
-                }
-            };
-
-            // This is a parameter for the outer lambda, that is stored in the closure for the func.
-            var context = Expression.Parameter(typeof(T), "context");
-
-            // These are passed at every host func invocation.
-            var machine = Expression.Parameter(typeof(Machine), "machine");
-            var frame = Expression.Parameter(typeof(Frame), "frame");
-
-            // Parameter Processing
-            var argsCount = 0;
-            var parameters = new List<Expression>();
-            foreach (var param in method.GetParameters())
-            {
-                if (param.ParameterType == typeof(Machine))
-                {
-                    parameters.Add(machine);
-                }
-                else if (param.ParameterType == typeof(Frame))
-                {
-                    parameters.Add(frame);
-                }
-                else
-                {
-                    parameters.Add(
-                        Expression.Call(
-                            Expression.ArrayIndex(
-                                Expression.PropertyOrField(frame, nameof(Frame.Locals)),
-                                Expression.Constant(argsCount)
-                            ),
-                            typeof(Value)
-                                .GetMethod(nameof(Value.As))
-                                .MakeGenericMethod(param.ParameterType)
-                        )
-                    );
-
-                    apiData
-                        .Data
-                        .Parameters
-                        .Add(
-                            new Parameter
-                            {
-                                Name = param.Name,
-                                Type = ValueTypeFor(param.ParameterType),
-                                CSType = param.ParameterType.GetNiceName()
-                            }
-                        );
-                    apiData.Data.ParameterValueTypes.Add(ValueTypeFor(param.ParameterType));
-
-                    argsCount++;
-                }
-            }
-
-            // The actual inner call.
-            var result = Expression.Call(method.IsStatic ? null : context, method, parameters);
-
-            var returnsCount = 0;
-            if (method.ReturnType != typeof(void))
-            {
-                returnsCount++;
-                // Process return value.
-                result = Expression.Call(
-                    frame,
-                    typeof(Frame)
-                        .GetMethods()
-                        .First(m => m.Name == nameof(Frame.Push) && m.IsGenericMethod)
-                        .MakeGenericMethod(method.ReturnType),
-                    result
-                );
-
-                apiData
-                    .Data
-                    .Returns
-                    .Add(
-                        new Parameter
-                        {
-                            Type = ValueTypeFor(method.ReturnType),
-                            CSType = method.ReturnType.GetNiceName()
-                        }
-                    );
-                apiData.Data.ReturnValueTypes.Add(ValueTypeFor(method.ReturnType));
-
-                // TODO: Add the ability to process value tuple based return values.
-            }
-
-            // This is the invoked callsite. To improve per-call performance, optimize the expression structure going into this compilation.
-            var callImpl = Expression.Lambda<Action<Machine, Frame>>(result, machine, frame);
-
-            var funcCtor = Expression.New(
-                typeof(HostFunc).GetConstructors().First(),
-                Expression.Constant(apiData.Data.Module),
-                Expression.Constant(apiData.Data.Name),
-                Expression.Constant(
-                    new FuncType(
-                        apiData.Data.ParameterValueTypes.ToArray(),
-                        apiData.Data.ReturnValueTypes.ToArray()
-                    )
-                ),
-                Expression.New(
-                    typeof(ActionHostProxy).GetConstructors().First(),
-                    callImpl,
-                    Expression.Constant(argsCount),
-                    Expression.Constant(returnsCount)
-                )
-            );
-
-            var lambda = Expression.Lambda<Func<T, HostFunc>>(funcCtor, context);
-            apiData.Function = lambda.Compile();
-            return apiData;
         }
 
         public List<HostFunc> Functions { get; }
-        public List<Api> ApiData { get; }
+        public List<ApiFunc> ApiData { get; }
 
         public HostFunc this[string name]
         {
             get => Functions.First(f => f.Name == name);
         }
 
-        public Api ApiDataFor(string name)
+        public ApiFunc ApiDataFor(string name)
         {
             return ApiData.First(f => f.Name == name);
         }
 
         public ReflectedModule(T self)
         {
-            Functions = ReflectedFuncs.Select(f => f.Function(self)).ToList();
+            Functions = ReflectedFuncs.Select(f => f.FuncFactory(self)).ToList();
             ApiData = ReflectedFuncs.Select(f => f.Data).ToList();
         }
     }
