+++ dergwasm_mod/Dergwasm/ResoniteEnv.cs
@@ -5,9 +5,6 @@ using Derg.Resonite;
 using Derg.Wasm;
 using Elements.Core;
 using FrooxEngine;
-using ProtoFlux.Core;
-using ProtoFlux.Runtimes.Execution.Nodes.FrooxEngine.Slots;
-using static FrooxEngine.SessionControlDialog;
 
 namespace Derg
 {
