+++ dergwasm_mod/DergwasmTests/instructions/ControlInstructionTests.cs
@@ -1,4 +1,5 @@
 ﻿using Derg;
+using Derg.Modules;
 using Xunit;
 
 namespace DergwasmTests.instructions
@@ -699,7 +700,8 @@ namespace DergwasmTests.instructions
             //
             // Func 14 (= idx 4): host func
             machine.SetProgram(0, I32Const(10), I32Const(20), Call(4), Nop());
-            machine.SetHostFuncAt(14, new ReturningHostProxy<int, int, int>((f, a, b) => a - b));
+            var hostFunc = ModuleReflector.ReflectHostFunc("test", "env", new System.Func<int, int, int>((a, b) => a - b));
+            machine.SetHostFuncAt(14, hostFunc.FuncFactory().Proxy);
 
             machine.Step(2);
             machine.Step();
